反馈数据增加版本号 优化更新操作,增加手动更新方式

+++ admin/inc/lic.php
@@ -51,30 +51,34 @@ if($config['title']&&$config['enddate']&&$config['date']&&$config['vip']&&$confi
             $sql="update config set vip='".base64_encode($result->vip)."',domain='".base64_encode($result->domain)."',templates='".base64_encode($result->templates)."',enddate='".base64_encode($result->enddate)."',date='".base64_encode(mt_rand(strtotime(date('Y-m-d',strtotime("+1 day"))),strtotime(date('Y-m-d',strtotime("+2 day")))))."',link=0 limit 1";
             $mysqli->query($sql);
             //验证成功,发送当前数据到服务器:域名个数,昨日蜘蛛数量
-            //获取域名数量
-            $domain_num=data_num("domains");
-            $spider_num=data_num('spider','',date('Y-m-d',time()-1*24*3600));
+            if($mysqli->affected_rows) {
+                //获取域名数量
+                $domain_num = data_num("domains");
+                $spider_num = data_num('spider', '', date('Y-m-d', time() - 1 * 24 * 3600));
 
-            $post_data['act']="data";
-            $post_data['domain_num']=$domain_num;
-            $post_data['spider_num']=$spider_num;
-            $request=request_post($post_data);
-            //file_get_contents("http://vip.alizhizhuchi.top/index.php?act=data&domain=".$yuming."&domain_num=".$domain_num."&spider_num=".$spider_num);//修改服务器域名
-            //域名限制
-            $vip_domain_num=$result->domain;
-            $domain_num=$mysqli->query("select count(*) as count from domains")->fetch_object()->count;
-            if($domain_num>$vip_domain_num){
-                //如果超出限制,删除多余
-                $del_num=$domain_num-$vip_domain_num;
-                $mysqli->query("delete from domains order by id desc limit ".$del_num);
-            }
-            //模板限制
-            $vip_templates_num=$result->templates;
-            $templates_num=$mysqli->query("select count(*) as count from templates where ok=1")->fetch_object()->count;
-            if($templates_num>$vip_templates_num){
-                //如果超出限制,禁用多余
-                $del_num=$templates_num-$vip_templates_num;
-                $mysqli->query("update templates set ok=0 where ok=1 order by id desc limit ".$del_num);
+                $post_data['act'] = "data";
+                $post_data['domain_num'] = $domain_num;
+                $post_data['spider_num'] = $spider_num;
+                $post_data['ver_title']=$config['ver'];
+                $request = request_post($post_data);
+                //file_get_contents("http://vip.alizhizhuchi.top/index.php?act=data&domain=".$yuming."&domain_num=".$domain_num."&spider_num=".$spider_num);//修改服务器域名
+
+                //域名限制
+                $vip_domain_num = $result->domain;
+                $domain_num = $mysqli->query("select count(*) as count from domains")->fetch_object()->count;
+                if ($domain_num > $vip_domain_num) {
+                    //如果超出限制,删除多余
+                    $del_num = $domain_num - $vip_domain_num;
+                    $mysqli->query("delete from domains order by id desc limit " . $del_num);
+                }
+                //模板限制
+                $vip_templates_num = $result->templates;
+                $templates_num = $mysqli->query("select count(*) as count from templates where ok=1")->fetch_object()->count;
+                if ($templates_num > $vip_templates_num) {
+                    //如果超出限制,禁用多余
+                    $del_num = $templates_num - $vip_templates_num;
+                    $mysqli->query("update templates set ok=0 where ok=1 order by id desc limit " . $del_num);
+                }
             }
         }else{
             $link=$mysqli->query("select link from config limit 1")->fetch_object()->link;
